separated the view controllers
