+++ Yona/Yona/Others/Constants.swift
@@ -86,7 +86,7 @@ struct YonaConstants {
         static let errorAddBuddyUserExists = "error.user.exists.created.on.buddy.request"
         static let errorUserNotFound = "error.user.not.found.id"
         static let tooManyOTPAttemps = "error.too.many.wrong.attempts"
-        static let tooManyResendOTPAttemps = "error.mobile.number.confirmation.code.too.many.failed.attempts"
+        static let tooManyFailedConfirmOTPAttemps = "error.mobile.number.confirmation.code.too.many.failed.attempts"
         static let tooManyPinResetAttemps = "error.pin.reset.request.confirmation.code.too.many.failed.attempts"
         static let cannotAddSecondGoalOnSameCategory = "error.goal.cannot.add.second.on.activity.category"
         static let cannotRemoveMandatoryGoal = "error.goal.cannot.remove.mandatory"
