+++ Yona/Yona/Others/SMSValidPasscodeLogin/SMSValidation/ConfirmMobileValidationVC.swift
@@ -1,0 +1,100 @@
+//
+//  SMSValidationViewController.swift
+//  Yona
+//
+//  Created by Chandan on 01/04/16.
+//  Copyright © 2016 Yona. All rights reserved.
+//
+
+import UIKit
+
+class ConfirmMobileValidationVC: ValidationMasterView {
+    @IBOutlet var resendOTPConfirmCodeButton: UIButton!
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        //Nav bar Back button.
+        self.navigationItem.hidesBackButton = true
+        
+        let viewWidth = self.view.frame.size.width
+        let customView=UIView(frame: CGRectMake(0, 0, (viewWidth-60)/2, 2))
+        customView.backgroundColor=UIColor.yiDarkishPinkColor()
+        
+        self.progressView.addSubview(customView)
+        
+        self.navigationController?.setNavigationBarHidden(false, animated: false)
+        
+        #if DEBUG
+            print ("pincode is \(YonaConstants.testKeys.otpTestCode)")
+        #endif
+    }
+    override func viewWillAppear(animated: Bool) {
+        super.viewWillAppear(animated)
+        setBackgroundColour()
+        
+        self.codeInputView.delegate = self
+        self.codeInputView.secure = true
+        codeView.addSubview(self.codeInputView)
+        
+        //keyboard functions
+        let notificationCenter = NSNotificationCenter.defaultCenter()
+        notificationCenter.addObserver(self, selector: Selector.keyboardWasShown, name: UIKeyboardDidShowNotification, object: nil)
+        notificationCenter.addObserver(self, selector: Selector.keyboardWillBeHidden, name: UIKeyboardWillHideNotification, object: nil)
+    }
+}
+
+extension ConfirmMobileValidationVC: CodeInputViewDelegate {
+    func codeInputView(codeInputView: CodeInputView, didFinishWithCode code: String) {
+
+        let body =
+            [
+                YonaConstants.jsonKeys.bodyCode: code
+            ]
+        Loader.Show()
+        UserRequestManager.sharedInstance.confirmMobileNumber(body) { success, message, serverCode in
+            Loader.Hide()
+            if (success) {
+                self.codeInputView.resignFirstResponder()
+                //Update flag
+                
+                setViewControllerToDisplay(ViewControllerTypeString.passcode, key: YonaConstants.nsUserDefaultsKeys.screenToDisplay)
+                self.performSegueWithIdentifier(R.segue.confirmMobileValidationVC.transToSetPincode, sender: self)
+
+                self.codeInputView.clear()
+                
+            } else {
+                self.checkCodeMessageShowAlert(message, serverMessageCode: serverCode, codeInputView: codeInputView)
+                self.codeInputView.clear()
+            }
+        }
+        
+    }
+}
+
+
+
+
+extension ConfirmMobileValidationVC: KeyboardProtocol {
+    func keyboardWasShown (notification: NSNotification) {
+        
+        if let activeField = self.resendOTPConfirmCodeButton, keyboardSize = (notification.userInfo?[UIKeyboardFrameBeginUserInfoKey] as? NSValue)?.CGRectValue() {
+            let contentInsets = UIEdgeInsets(top: 0.0, left: 0.0, bottom: keyboardSize.height, right: 0.0)
+            self.scrollView.contentInset = contentInsets
+            self.scrollView.scrollIndicatorInsets = contentInsets
+            var aRect = self.scrollView.bounds
+            aRect.size.height -= keyboardSize.size.height
+            if (!CGRectContainsPoint(aRect, activeField.frame.origin)) {
+                var frameToScrollTo = activeField.frame
+                frameToScrollTo.size.height += 30
+                self.scrollView.scrollRectToVisible(frameToScrollTo, animated: true)
+            }
+        }
+    }
+    
+    func keyboardWillBeHidden(notification: NSNotification) {
+        let contentInsets = UIEdgeInsetsZero
+        self.scrollView.contentInset = contentInsets
+        self.scrollView.scrollIndicatorInsets = contentInsets
+        
+    }
+}
