+++ Yona/Yona/UserRequestManager.swift
@@ -1,4 +1,4 @@
-//
+ //
 //  UserRequestManager.swift
 //  Yona
 //
